Refactor state SSH wrapper integration tests
